+++ ijele/components/Dashboard/tripTab.tsx
@@ -2,7 +2,20 @@
 import { ChangeEvent, useState } from 'react'
 import { Dialog, DialogBackdrop, DialogPanel, DialogTitle } from '@headlessui/react'
 import Image from 'next/image'
-import { NewTripForm } from './new-trip'
+import { NewTripForm } from './tripForm'
+
+
+interface NewTripProps {
+    tripDetails: {
+        title: string,
+        description: string,
+    };
+    
+}
+interface CardArray{
+    tripDetailsArr: NewTripProps[],
+    handleInputChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
 
 function groupMates() {
     return (
@@ -36,13 +49,41 @@ function groupMates() {
     )
 }
 
+function buildTripCard() {
+    const cards: JSX.Element[] = [];
+    return (
+        <div className="card card-side bg-base-100 shadow-xl">
+            <figure>
+                <Image src='/Images/hotel/results/resort1.jpg' width={300} height={300} alt={'trip image'} />
+            </figure>
+            <div className="card-body">
+                <div className="card-actions flex">
+                    <h2 className="card-title text-ijele_teal text-xl">Atlanta Trip!</h2>
+                    <button disabled> <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" height="5%" width="5%" className='fill-ijele_sand'><path d="M441 58.9L453.1 71c9.4 9.4 9.4 24.6 0 33.9L424 134.1 377.9 88 407 58.9c9.4-9.4 24.6-9.4 33.9 0zM209.8 256.2L344 121.9 390.1 168 255.8 302.2c-2.9 2.9-6.5 5-10.4 6.1l-58.5 16.7 16.7-58.5c1.1-3.9 3.2-7.5 6.1-10.4zM373.1 25L175.8 222.2c-8.7 8.7-15 19.4-18.3 31.1l-28.6 100c-2.4 8.4-.1 17.4 6.1 23.6s15.2 8.5 23.6 6.1l100-28.6c11.8-3.4 22.5-9.7 31.1-18.3L487 138.9c28.1-28.1 28.1-73.7 0-101.8L474.9 25C446.8-3.1 401.2-3.1 373.1 25zM88 64C39.4 64 0 103.4 0 152L0 424c0 48.6 39.4 88 88 88l272 0c48.6 0 88-39.4 88-88l0-112c0-13.3-10.7-24-24-24s-24 10.7-24 24l0 112c0 22.1-17.9 40-40 40L88 464c-22.1 0-40-17.9-40-40l0-272c0-22.1 17.9-40 40-40l112 0c13.3 0 24-10.7 24-24s-10.7-24-24-24L88 64z" /></svg></button>
+                </div>
+                <div className='justify-start text-start flex'>
+                    <label htmlFor="location" className='mr-3 text-gray-400'>Location</label>
+                    <p className='font-kite_one'>Atlanta</p>
+                </div>
+                <div className='justify-start text-start flex'>
+                    <label htmlFor="description" className='mr-3 text-gray-400'>Description</label>
+                    <p className='font-kite_one'>Get the Girls out the groupchat and into our choice get away!</p>
+                </div>
+                <div className='text-start items-center '>
+                    <label htmlFor="Travlers" className='font-kite_one text-gray-400'>Travelers</label>
+                    {groupMates()}
+                </div>
+            </div>
+        </div>
+    )
+}
 export const UserTrips = () => {
 
     const [open, setOpen] = useState(false)
+    const [cardObject, setCardObjects] = useState<CardArray>()
     const [newTrip, setNewTrip] = useState({
         title: "",
         description: "",
-        travelers: []
     })
 
     const handleTripInputChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -51,10 +92,11 @@ export const UserTrips = () => {
             ...prevDetails,
             [name]: value,
         }));
+
+        
     };
-    const newTripCard = async (e: React.FormEvent) => {
-        e.preventDefault();
-        console.log('create new card')
+    const newTripCard = (cardCreated: boolean) => {
+        setObjectReturned()
     }
 
     return (
@@ -87,7 +129,7 @@ export const UserTrips = () => {
                                                 <p className="text-sm text-gray-500">
                                                     Please Fill the following information completely!
                                                 </p>
-                                                <form method="post">
+                                                <form>
                                                     <NewTripForm tripDetails={newTrip} handleInputChange={handleTripInputChange} />
                                                 </form>
                                             </div>
@@ -105,11 +147,11 @@ export const UserTrips = () => {
                                     <button
                                         type="button"
                                         data-autofocus
-                                        onClick={() =>{
+                                        onClick={() => {
                                             newTripCard
                                             setOpen(false)
 
-                                        }} 
+                                        }}
                                         className="mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50 sm:mt-0 sm:w-auto"
                                     >
                                         Save
@@ -121,29 +163,7 @@ export const UserTrips = () => {
                 </Dialog>
 
                 <div className='mt-8 max-h-sm'>
-                    <div className="card card-side bg-base-100 shadow-xl">
-                        <figure>
-                            <Image src='/Images/hotel/results/resort1.jpg' width={300} height={300} alt={'trip image'} />
-                        </figure>
-                        <div className="card-body">
-                            <div className="card-actions flex">
-                                <h2 className="card-title text-ijele_teal text-xl">Atlanta Trip!</h2>
-                                <button disabled> <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" height="5%" width="5%" className='fill-ijele_sand'><path d="M441 58.9L453.1 71c9.4 9.4 9.4 24.6 0 33.9L424 134.1 377.9 88 407 58.9c9.4-9.4 24.6-9.4 33.9 0zM209.8 256.2L344 121.9 390.1 168 255.8 302.2c-2.9 2.9-6.5 5-10.4 6.1l-58.5 16.7 16.7-58.5c1.1-3.9 3.2-7.5 6.1-10.4zM373.1 25L175.8 222.2c-8.7 8.7-15 19.4-18.3 31.1l-28.6 100c-2.4 8.4-.1 17.4 6.1 23.6s15.2 8.5 23.6 6.1l100-28.6c11.8-3.4 22.5-9.7 31.1-18.3L487 138.9c28.1-28.1 28.1-73.7 0-101.8L474.9 25C446.8-3.1 401.2-3.1 373.1 25zM88 64C39.4 64 0 103.4 0 152L0 424c0 48.6 39.4 88 88 88l272 0c48.6 0 88-39.4 88-88l0-112c0-13.3-10.7-24-24-24s-24 10.7-24 24l0 112c0 22.1-17.9 40-40 40L88 464c-22.1 0-40-17.9-40-40l0-272c0-22.1 17.9-40 40-40l112 0c13.3 0 24-10.7 24-24s-10.7-24-24-24L88 64z" /></svg></button>
-                            </div>
-                            <div className='justify-start text-start flex'>
-                                <label htmlFor="location" className='mr-3 text-gray-400'>Location</label>
-                                <p className='font-kite_one'>Atlanta</p>
-                            </div>
-                            <div className='justify-start text-start flex'>
-                                <label htmlFor="description" className='mr-3 text-gray-400'>Description</label>
-                                <p className='font-kite_one'>Get the Girls out the groupchat and into our choice get away!</p>
-                            </div>
-                            <div className='text-start items-center '>
-                                <label htmlFor="Travlers" className='font-kite_one text-gray-400'>Travelers</label>
-                                {groupMates()}
-                            </div>
-                        </div>
-                    </div>
+                    {buildTripCard()}
                 </div>
 
             </div>
